Fix teachers API import path

+++ src/app/api/test-db/route.ts
@@ -5,7 +5,7 @@ import { apiResponse, apiError } from '@/lib/api/utils'
 export async function GET(request: NextRequest) {
   try {
     // Test simple database connection
-    const result = await db.$queryRaw`SELECT COUNT(*) as count FROM events`
+    const result = await db.$queryRaw`SELECT COUNT(*)::int as count FROM events`
     
     return apiResponse({
       message: 'Database connection successful',
